Moving app launch time tests to their own class and resetting static state for them
